refactor(ui): share cdp-scripts test helpers and trim loader
Extract the repeated `cdpDoc` / `resetClickState` / `resetHoverState`
setup from the six cdp-scripts vitest suites into a shared
`test-helpers.ts`. Drop the unused `rawCommon` export and stale
eslint-disable hints from `loader.ts`.
