+++ ui/src/cdp-scripts/loader.ts
@@ -1,10 +1,6 @@
-// Helper that loads the CDP injected-JS scripts from the Rust crate so we
-// can test their observable behavior in jsdom.
-//
-// The scripts live in `crates/clickweave-core/src/walkthrough/cdp_scripts/`
-// and are assembled into a single evaluated function at Rust compile time
-// via `concat!(include_str!(...), ...)`.  To mirror that at test time we
-// import each file with Vite's `?raw` suffix and compose them the same way.
+// Load each CDP injected-JS script via Vite's `?raw` import and compose it
+// the same way Rust does at compile time, so vitest runs the exact bytes
+// that ship to the browser.
 
 import commonSrc from "../../../crates/clickweave-core/src/walkthrough/cdp_scripts/common.js?raw";
 import clickListenerSrc from "../../../crates/clickweave-core/src/walkthrough/cdp_scripts/click_listener.js?raw";
@@ -14,16 +10,12 @@ import retrieveClickSrc from "../../../crates/clickweave-core/src/walkthrough/cd
 import retrieveHoversSrc from "../../../crates/clickweave-core/src/walkthrough/cdp_scripts/retrieve_hovers.js?raw";
 import stopHoverSrc from "../../../crates/clickweave-core/src/walkthrough/cdp_scripts/stop_hover.js?raw";
 
-/** Compose `common.js` + a listener body into a callable arrow function. */
 function composeWithCommon(bodySrc: string): () => unknown {
     const src = `() => {\n${commonSrc}${bodySrc}}`;
-    // eslint-disable-next-line no-new-func
     return new Function(`return (${src});`)() as () => unknown;
 }
 
-/** Wrap a standalone `() => {...}` script as a callable arrow function. */
 function composeStandalone(src: string): () => unknown {
-    // eslint-disable-next-line no-new-func
     return new Function(`return (${src});`)() as () => unknown;
 }
 
@@ -51,5 +43,3 @@ export function loadRetrieveHovers(): () => unknown[] {
 export function loadStopHover(): () => void {
     return composeStandalone(stopHoverSrc) as () => void;
 }
-
-export const rawCommon = commonSrc;
